Init Dart 3 records support
