+++ packages/auto_mappr/build.yaml
@@ -4,6 +4,8 @@ targets:
       :auto_mappr:
         enabled: true
         generate_for:
+          include: # TODO
+            - test/integration/fixture/record.dart
           exclude:
             - test/integration/error_fixture/**
         options:
