+++ packages/auto_mappr/lib/src/builder/auto_mappr_builder.dart
@@ -9,6 +9,7 @@ import 'package:auto_mappr/src/models/type_mapping.dart';
 import 'package:built_collection/built_collection.dart';
 import 'package:code_builder/code_builder.dart';
 
+/// Entrypoint for mappr class generation.
 class AutoMapprBuilder {
   final AutoMapprConfig config;
   final ClassElement mapperClassElement;
