+++ packages/auto_mappr/lib/src/builder/map_bodies/map_bodies.dart
@@ -1,0 +1,2 @@
+export 'class_body_builder.dart';
+export 'enum_body_builder.dart';
