+++ packages/auto_mappr/pubspec.yaml
@@ -12,7 +12,7 @@ environment:
 
 dependencies:
   # TODO: Support analyzer same way as json_serializable https://github.com/google/json_serializable.dart/pull/1333
-  analyzer: ">=5.4.0 <7.0.0"
+  analyzer: ^6.0.0
   auto_mappr_annotation: ^1.2.0
   build: ^2.3.1
   built_collection: ^5.1.1
